+++ src/builder/index.js
@@ -80,7 +80,7 @@ class Builder {
 
     // 2. Modules integration
     if (this.options.debug) {
-      _.forOwn(ms.moduleCollection, (value) => {
+      Object.values(ms.moduleCollection).forEach((value) => {
         let relPath = path.relative(this._coreDirname, value.filename + '.json');
         let absPath = path.join(this._metaDirname, relPath);
         let str = JSON.stringify(value.parsed, null, 2);