+++ src/xlsx-connector/index.js
@@ -33,16 +33,17 @@ function convertExcelSync(src, dst = null, options = {}){
 
 function _toDeepPaths(o){
   let output = {};
-  _.each((o), (value, key) => {
-    let searchArray = /^(.+)\[\]$/;
-    if (searchArray.test(key)) { // checking if path looks like this "one.two.three[]"
-      let keyPart = key.match(searchArray)[1];
-      let valuesPart = value.toString().split(';').filter((x) => _.trim(x) !== '');
-      _.set(output, keyPart, valuesPart);
-    } else {
-      _.set(output, key, value);
-    }
-  });
+  Object.entries(o)
+    .forEach(([key, value]) => {
+      let searchArray = /^(.+)\[\]$/;
+      if (searchArray.test(key)) { // checking if path looks like this "one.two.three[]"
+        let keyPart = key.match(searchArray)[1];
+        let valuesPart = value.toString().split(';').filter((x) => _.trim(x) !== '');
+        _.set(output, keyPart, valuesPart);
+      } else {
+        _.set(output, key, value);
+      }
+    });
 
   return output;
 }